+++ lib/hl_tcl/pkgIndex.tcl
@@ -1,5 +1,5 @@
 
-package ifneeded hl_tcl 1.0.1 [list source [file join $dir hl_tcl.tcl]]
+package ifneeded hl_tcl 1.0.2 [list source [file join $dir hl_tcl.tcl]]
 
 # short intro (for Ruff! docs generator)